refactor: globalstyle, theme 위치 변경
