+++ FrontEnd/src/App.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Global, ThemeProvider } from '@emotion/react';
 
-import { globalStyle, theme } from './App.style';
+import { globalStyle } from './components/@styled/globalStyle';
+import { theme } from './constants/theme';
 import Login from './pages/Login/Login';
 import Posts from './pages/Posts';
 import SignUp from './pages/SignUp/SignUp';
