+++ FrontEnd/src/constants/theme.ts
@@ -1,0 +1,18 @@
+export const theme = {
+  color: {
+    appBackgroundColor: '#ffffff',
+    primaryColor: '#fc3465',
+    darkerSecondaryColor: '#ffdede',
+    secondaryColor: '#fff0f0',
+    tertiaryColor: '#6d6d6d',
+    textColor: '#3d3d3d',
+    lighterTextColor: '#959595',
+    postBackgroundColor: '#f2f2f2',
+    borderColor: '#cfcfcf',
+    darkBorderColor: '#9f9f9f',
+    tagItemColor: '#ffc1c1',
+    separatorColor: '#a4a4a45b',
+    yellow: '#ffe812',
+    white: '#ffffff',
+  },
+};
